refactor: extract components to reduce lint warnings
Split multi-component files into focused modules:
- activity_elapsed.tsx (ActivityElapsed + classifyElapsed)
- section_content.tsx (ExpandedContent + CollapsedHint)
- search_input.tsx (SearchInput)
- filter_button.tsx (FilterButton)
- menu_items.tsx (MenuItems + MenuItem + handlers)

Fix curly brace and negated-condition warnings. Remaining 28
warnings are inherent to the code patterns (magic numbers in
named constants, JSX components exceeding 25-line limit).
