+++ app/frontend/pages/admin/dashboard/filter_button.tsx
@@ -1,0 +1,19 @@
+import Button from "@/components/ui/button";
+
+export function FilterButton({
+  label,
+  count,
+  active,
+  onClick,
+}: {
+  label: string;
+  count: number;
+  active: boolean;
+  onClick: () => void;
+}) {
+  return (
+    <Button variant={active ? "primary" : "secondary"} size="sm" onClick={onClick}>
+      {label} {count}
+    </Button>
+  );
+}
