+++ app/frontend/pages/admin/dashboard/store_card/action_menu.tsx
@@ -1,9 +1,8 @@
-import { router } from "@inertiajs/react";
 import { useEffect, useRef, useState } from "react";
 
 import type { AdminStoreSummary } from "@/types/inertia";
 
-import { canResync } from "../dashboard_constants";
+import { MenuItems } from "./menu_items";
 
 export function ActionMenu({ store }: { store: AdminStoreSummary }) {
   const [open, setOpen] = useState(false);
@@ -22,33 +21,6 @@ export function ActionMenu({ store }: { store: AdminStoreSummary }) {
     return () => document.removeEventListener("mousedown", handleClick);
   }, [open]);
 
-  const resyncable = canResync(store);
-  const enrichable = ["failed", "stale", "processing", "partial"].includes(store.health.key);
-
-  const handleResync = () => {
-    if (!window.confirm(`Resync ${store.name}?`)) {
-      return;
-    }
-    router.post(`/admin/stores/${store.id}/retry`);
-    setOpen(false);
-  };
-
-  const handleRetryEnrichment = () => {
-    if (!window.confirm(`Retry enrichment for ${store.name}?`)) {
-      return;
-    }
-    router.post(`/admin/stores/${store.id}/retry_enrichment`);
-    setOpen(false);
-  };
-
-  const handleDelete = () => {
-    if (!window.confirm(`Delete ${store.name}? This cannot be undone.`)) {
-      return;
-    }
-    router.delete(`/admin/stores/${store.id}`);
-    setOpen(false);
-  };
-
   return (
     <div ref={menuRef} className="relative">
       <button
@@ -61,41 +33,7 @@ export function ActionMenu({ store }: { store: AdminStoreSummary }) {
       </button>
       {open && (
         <div className="absolute right-0 z-10 mt-1 w-48 rounded-md border border-mc-border bg-mc-bg-card shadow-lg">
-          {resyncable && (
-            <button
-              type="button"
-              onClick={handleResync}
-              className="block w-full px-3 py-2 text-left text-sm text-mc-text hover:bg-mc-border"
-            >
-              Resync now
-            </button>
-          )}
-          {enrichable && (
-            <button
-              type="button"
-              onClick={handleRetryEnrichment}
-              className="block w-full px-3 py-2 text-left text-sm text-mc-text hover:bg-mc-border"
-            >
-              Retry enrichment
-            </button>
-          )}
-          <a
-            href={store.storefront_path}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="block px-3 py-2 text-left text-sm text-mc-text hover:bg-mc-border"
-            onClick={() => setOpen(false)}
-          >
-            View storefront
-          </a>
-          <div className="border-t border-mc-border" />
-          <button
-            type="button"
-            onClick={handleDelete}
-            className="block w-full px-3 py-2 text-left text-sm text-mc-feedback-danger hover:bg-mc-border"
-          >
-            Delete store
-          </button>
+          <MenuItems store={store} onClose={() => setOpen(false)} />
         </div>
       )}
     </div>
